+++ src/components/Form.js
@@ -1,0 +1,8 @@
+import React from 'react';
+
+
+const Form = () => {
+    return <div></div>
+}
+
+export default Form;
